requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
joblib==1.4.2
numpy==2.1.0
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
scikit-learn==1.5.1
scipy==1.14.1
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
